Some cuts here and there...
a) The introduction to Section 2 has been reduced to the bare minimum.
The text mostly overlaps with Section 1 (Introduction), so no real
loss of information.

b) Removed the description of the tools from Section 2. There was a
LaTeX comment from AP saying that we need the description of the tools
for Table 1 and Table 2, but I think it is fine like it is (i.e.
describing the tools only in Section 4). Table 1 and Table 2 appear in
the paper when Section 4 has already started, so it looks fine.

c) While we are talking about Table 1 and Table 2 (use case
scenarios), I am not sure these tables must contain information on how
to find the information (e.g. "The teacher can consult the CD tool and
see..."). Although this is nice to have, it also overlaps with Section
4.5 (example of use of TA tools), makes the tables very big, and I
found it a bit distracting when I read the tables.

d) Removed the list of important indicators from Section 4.1. It is
not crucial for this paper, it overlaps with the TA-1 paper, and it
uses a lot of space. Added an additional reference to the TA-1 paper.

e) Shortened the first paragraph of the discussion (related to
formative evaluation). Also shortened the paragraph starting "All
these results point to...".

I think we can reduce the formative evaluation to half its size, but
this would take time. Probably better to leave it as it is for now,
and wait for the reviewers to complain (but hopefully not reject the
paper).